requirements.txt
pyyaml
requests
xmltodict
